+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Add option for delaying the webhook. Give the frontend time to persist.
 
+### FIX
+
+- fix OrderNumber-Gaps in PayPalExpress-Orders
+- use cancel-Process also for PayPalExpress
+- Introduce orderModel->cancelPayPalOrder. It takes over a large part of the tasks of the method PaymentService->removeTemporaryOrder
+
 ## [3.5.3] - 2026-02-02
 
 ### FIX